git-svn-id: https://smlnj-gforge.cs.uchicago.edu/svn/manticore/branches/proxy-cont@4738 ac4fbcad-d21d-0410-b760-c7b1d3380187

+++ src/lib/parallel-rt/config/gen-runtime-constants.c
@@ -24,7 +24,7 @@ static int LogBufSz = LOGBUF_SZ;
 
 /* print the definition of a symbol */
 #define PR_DEFINE(symb, val)							\
-	printf("    val " #symb " : IntInf.int = %#0llx\n", (uint64_t)val)
+	printf("    val " #symb " : IntInf.int = %#0llx\n", (unsigned long long int)val)
 
 /* print a value definition and record it in the CRC buffer */
 #define PR_VALUE(tag, var, value)						\
